Statlab_IA_Project/data-cleaning-service/requirements.txt
kafka-python==2.0.2
minio==7.2.5
pandas==2.2.2
python-dotenv==1.0.1
